requirements.txt
flask
gunicorn
boto3
faiss-cpu
flask-jwt-extended
passlib
python-dotenv
openai
sentence-transformers